requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
blinker==1.8.2
boto3==1.34.153
botocore==1.34.153
Brotli==1.1.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==5.0.7
django-cachalot==2.6.3
django-cache-machine==1.2.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-rest-swagger==2.2.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
locust==2.30.0
MarkupSafe==2.1.5
moviepy==1.0.3
msgpack==1.0.8
numpy==2.0.1
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
setuptools==72.1.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
